move all react files in script folder
