+++ direct/_scripts/components/FAQ.js
