+++ direct/_scripts/components/Privacy.js
