+++ direct/_scripts/components/computeDetour.js
