+++ direct/_scripts/index.html
@@ -182,6 +182,9 @@
 	<script type="module" src="build/bundle.js"></script>
 </head>
 
-<body></body>
+<body>
+	<div class='react-component'></div>
+
+</body>
 
 </html>
