--- instances/valberg.yaml
@@ -1,11 +1,0 @@
-nom: Azur.comobi
-sous-titre: 'Circulez autour de Valberg en covoiturage !'
-description: Simple, gratuit, solidaire, azur.comobi est un service de covoiturage de proximité qui met en relation des conducteurs avec des personnes souhaitant partager leur trajet parce qu’ils n’ont pas de véhicule, permis ou parce qu’il préfèrent économiser leurs trajets !
-slogan: 'Pas de transports en communs ? votre territoire vous propose une solution facile !'
-formulaire: https://docs.google.com/forms/d/e/1FAIpQLSeInZJD79GJ-qg3Tkjz47uqsVgsZVy5awevxvhB67jK8KGGJQ/viewform?usp=sf_link
-image: https://imgur.com/NjjlUdi.png
-logo: https://upload.wikimedia.org/wikipedia/fr/c/cf/Valberg_logo_2012.svg
-contact:
-  nom: Station Valberg - Maureen DJENAIHI
-  mel: mdjenaihi@valberg.com
-  tel: 04 93 23 24 25
